+++ config/app.php
@@ -10,6 +10,7 @@ return [
 
     'changelog_url' => env('CHANGELOG_URL', 'https://changelog.datamermaid.org'),
     'changelog_path' => env('CHANGELOG_PATH', 'changelog/mermaid-changelog.json'),
+    'enable_sync_cloud' => env('ENABLE_SYNC_CLOUD', false),
 
     /*
     |--------------------------------------------------------------------------